feat: Add more unit tests
changes:
- unit tests fixed
- unit tests added

closes #6
