+++ tests/BB84.ExtensionsTests/StringExtensionsTests.cs
@@ -1,34 +1,10 @@
-﻿using System.Globalization;
-
-using BB84.Extensions;
+﻿using BB84.Extensions;
 
 namespace BB84.ExtensionsTests;
 
 [TestClass]
-public class StringExtensionsTests
+public sealed partial class StringExtensionsTests
 {
-	[TestMethod]
-	public void FormatInvariantSuccessTest()
-	{
-		DateTime dateTime = DateTime.Now;
-		string testString = @"Today is: {0}";
-
-		string result = testString.FormatInvariant(dateTime);
-
-		Assert.IsTrue(result.Contains(dateTime.ToString(CultureInfo.InvariantCulture)));
-	}
-
-	[TestMethod]
-	public void FormatSuccessTest()
-	{
-		DateTime dateTime = DateTime.Now;
-		string testString = @"Today is: {0}";
-
-		string result = testString.Format(CultureInfo.GetCultureInfo("de-DE"), dateTime);
-
-		Assert.IsTrue(result.Contains(dateTime.ToString(CultureInfo.GetCultureInfo("de-DE"))));
-	}
-
 	[TestMethod]
 	public void IsNullTest()
 	{
@@ -58,62 +34,11 @@ public class StringExtensionsTests
 	}
 
 	[TestMethod]
-	public void ToBase64Test()
-	{
-		string value = "UnitTest";
-
-		string result = value.ToBase64();
-
-		Assert.AreNotEqual(value, result);
-		Assert.AreEqual("VW5pdFRlc3Q=", result);
-	}
-
-	[TestMethod]
-	public void ToBase64EmptyTest()
-	{
-		string value = string.Empty;
-
-		string result = value.ToBase64();
-
-		Assert.AreEqual(value, result);
-	}
-
-	[TestMethod]
-	public void FromBase64Test()
-	{
-		string value = "VW5pdFRlc3Q=";
-
-		string result = value.FromBase64();
-
-		Assert.AreNotEqual(value, result);
-		Assert.AreEqual("UnitTest", result);
-	}
-
-	[TestMethod]
-	public void FromBase64EmptyTest()
-	{
-		string value = string.Empty;
-
-		string result = value.FromBase64();
-
-		Assert.AreEqual(value, result);
-	}
-
-	[TestMethod]
-	[ExpectedException(typeof(ArgumentException))]
-	public void FromBase64ExceptionTest()
-	{
-		string value = "%$#";
-
-		_ = value.FromBase64();
-	}
-
-	[TestMethod]
 	public void GetMd5Utf8Test()
 	{
 		string value = "UnitTest";
 
-		var result = value.GetMd5Utf8();
+		string result = value.GetMd5Utf8();
 
 		Assert.AreEqual("37ADC7DB47085615AF6389C9C50AF7B9", result);
 	}
@@ -123,7 +48,7 @@ public class StringExtensionsTests
 	{
 		string value = "UnitTest";
 
-		var result = value.GetMd5Unicode();
+		string result = value.GetMd5Unicode();
 
 		Assert.AreEqual("000C23A22CE00D7163E8FF10A23FCDC3", result);
 	}
@@ -133,7 +58,7 @@ public class StringExtensionsTests
 	{
 		string value = "UnitTest";
 
-		var result = value.GetMd5Ascii();
+		string result = value.GetMd5Ascii();
 
 		Assert.AreEqual("37ADC7DB47085615AF6389C9C50AF7B9", result);
 	}
